requirements.txt
requests==2.32.0
beautifulsoup4==4.11.1
selenium==4.15.1
webdriver-manager==3.8.5
pandas==2.2.3
sqlalchemy==2.0.0
mysql-connector-python==8.0.33
python-dotenv==0.21.0